requirements.txt
streamlit>=1.50.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
openai-whisper>=20230314
numpy>=1.24.0
scikit-learn>=1.3.0
librosa>=0.10.0
opencv-python>=4.7.0
